requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.50.0
pydantic>=2.9.0
chromadb>=0.5.0
python-docx>=1.1.0
reportlab>=4.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
httpx>=0.27.0
